requirements.txt
tqdm
numpy
matplotlib
pandas
scikit_learn
wget==3.2
pytest
torchvision
torchinfo
torchmetrics
einops
black>=22.6
coverage>=6
darglint>=1.8
isort>=5.10
pipreqs==0.4.11
pre-commit>=2.20
pytest-cov>=3
transformers
seaborn
SciencePlots
scikit-image
codecov
randomname
wandb
torch-lr-finder
loguru
fairlearn
